requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.27.0
apscheduler==3.10.4
pydantic==2.6.1
pydantic-settings==2.1.0
apify-client==1.8.1
